新增：测试用例

+++ download/src/main/java/com/example/newload/CallbackHandler.kt
@@ -25,7 +25,7 @@ class CallbackHandler private constructor() {
 
     fun notifyProgress(
         taskId: String,
-        progress: Int,
+        progress: Double,
         downloadedBytes: Long,
         totalBytes: Long,
         callback: DownloadCallback?
